Move tests to appropriate files
Now that everything is a bit more modular, the tests should reflect
that.
